'Quick Start' has new meaning, so replaceing with Inventory.
A couple SkyEye verbatims indicated those customers felt misled by the TOC title prefix of 'Quick Start', which nowadays has come to mean a brief procedure to create a simple example (something like that). Originally here 'Quick Start' was meant to mean a mention + link + short description of all the many subfeatures of In-Memory tables that the customer would at least need to be aware of before using this whole In-Memory feature.
So, replacing 'Quick Start:' with 'Inventory of'.
MightyPen = genemi.