Expanded definitions and updated for SQL Server 2017
Expanded definitions of "CodePage", "LCID", and "ComparisonStyle" properties so that they aren't a mystery to most people who don't know where to get those definitions. I copied the 4 bitmask values for "ComparisonStyle" from [DATABASEPROPERTYEX]( https://docs.microsoft.com/en-us/sql/t-sql/functions/databasepropertyex-transact-sql) but did not copy the "for example" portion as that seemed extraneous. But, if it is best that it be here for consistency then I can add it in.

Regarding the link used for "LCID": the page has the value in hex which requires the extra step of converting the INT value returned here, but that list is by far the most complete that I could find. In fact, I couldn't find any other list that included entries for values 0x00010407 through 0x00040411.

Regarding the addition to the "Version" info: I confirmed that info on SQL Server 2017 as **3** is returned for the new Japanese collations (i.e. the `Japanese_Bushu_Kakusu_140_` series and the `Japanese_XJIS_140_` series).