C3116: LOCBLOCKER - Missing content due to incorrect format
Hello, @BYHAM,

Localization team has reported source content issue that cause localized version to have blocked content. All < > which are incorrectly place should be removed. 

Please, help to check my proposed file change into the article and help to merge if you agree with fix. If not, please let me know either if you would like me to fix it in another way within this PR, if you prefer to fix it in another PR, or if I should close this PR as by-design. In case of using another PR, please, let me know of your PR number, so we can confirm and close this PR.

Many thanks in advance.