C6022: Incorrect struture for Date/Time Data Types
Hello, @edmacauley, 

There is a content issue that causes localized version to have broken/different format. The Date/Time Data Type seems to be incorrect.

Please, help to check my proposed file change and help to merge if you agree. If not, please, let me know either if you would like me to fix it in another way within this PR, if you prefer to fix it in another PR, or if I should close this PR as by-design. In case of using another PR, please, let me know of your PR number, so we can confirm and close this PR.

Many thanks in advance.