C6025: LOCBLOCKER - Missing ``` that blocking content for localization
Hello, @jeannt,

There is a source  issue that causes localizable text to be blocked from localization.  Missing ``` are blocking content for localization.
Please, help to check my proposed file change and help to merge if you agree with fix. In case of using another PR, please let me know of your PR number, so we can confirm and close this PR.

Many thanks in advance.