Change description of -i parameter
The description of the -i parameter was confusing. I have a default SQL instance, but the SQL Server 2017 Reporting Services silent installation created a name SSRS instance (name "SSRS"). Without specifying the -i parameter (because I am using a default SQL instance), I received a WMI error. When I specified "-i SSRS" in the command line, it worked. It set the database to use in the default SQL Server instance.