This topic omits returning data using Result Sets
This topic omits returning data using Result Sets, which is the most common way to return data from stored procedures, and the only way to return tabular data from stored procedures.  In addition the behavior of TSQL control flow w.r.t. returning results to clients is nowhere else documented, and this is a reasonable topic to do that.  Although a doc on TSQL batch execution in general (if one existed) would be a good place too.