Update k8s Azure disk link
The old link pointed to a deprecated Docker volume driver for Azure Files (SMB/CIFS)

Azure Disk support is included in Kubernetes upstream, and is a type of volume available, not adriver. It uses either page blobs in storage accounts (.vhd), or Managed Disks directly. The new link is the same as used the official Kubernetes docs to describe Azure Disk support, and has more info on advanced configuration options