Add recommendation to consider changing default value
The default value of 5 is now regarded as far too low by almost everyone working with the product. This reality should be recognized on this page. We typically start with 50 and often work up. I know that others start with 25 and work up. I don't know anyone who starts with the default value of 5. This guidance should be provided.