Included a stronger recommendation to leave this disabled
The current article discusses how to change the option but gives no hint on why it is disabled by default, and why it should normally remain that way. I've added content to explain that, and to suggest one of the limited situations where the option might need to be enabled.