M18114: Possible error: PolyBase Guide
Hello, @barbkess,

Translator has reported possible source content issue.  n this segment, the part relative to "PolyBase Guide" is confusing. We suppose that the feature name is just PolyBase, 

Please, help to check my proposed file change into the article and help to merge if you agree with fix. If not, please, let me know either if you would like me to fix it in another way within this PR, if you prefer to fix it in another PR, or if I should close this PR as by-design. In case of using another PR, please, let me know of your PR number, so we can confirm and close this PR. 

Many thanks in advance.