Fix display fail of huge wide table.
Ramu Konidena notified me that the huge table in this article fails to display because it is too wide. I agree the display was terrible.
So I combined the first three (of total four) columns into one column, using BR tags to delimit the logical fields that now share the first cell in each row.