technical inaccuracy in when uncommittable trans will be rolled back
If a transaction becomes uncommittable in a nested batch (using sp_executesql), the transaction will NOT be rolled back when this batch finishes. The transaction will be rolled back when the batch that called sp_executesql finishes. So it is incorrect to say "When a batch finishes running". I am not sure if "outermost batch" is the accepted terminology for this, but it seems reasonable.