Update misleading statement about TRUNCATE TABLE
The document says that TRUNCATE TABLE is non-logged, but that is incorrect and misleads the reader in a big way. It does not log the removal of individual rows, but the operation (removing pages) is captured in the transaction log. That is an important distinction because readers of this document may also be considering log shipping and will conclude that TRUNCATE TABLE is not compatible with that. Secondly, the document states that TRUNCATE TABLE does not fire triggers. While it does not fire DML triggers, which are used to detect changes for merge replication, it does fire DDL triggers.