In correct information existing powerpivot model
If the database exists and we try to restore we would get an error. “Object cannot be restored because the object already exists.  Please select a different database name.. There was Defect 743287:Remove Override checkbox in restore from PowerPivot filled to change the behavior. But the documentation suggest we can override an existing database which is not true.