Add info about automatic serial redo
An important detail (about which databases are automatically put into serial redo mode when more than 100 parallel redo threads are required) was released in the following blog post (near the end, under the heading "Parallel redo thread usage and redo model control"):

https://blogs.msdn.microsoft.com/sql_server_team/sql-server-20162017-availability-group-secondary-replica-redo-model-and-performance/

I think it would make sense for that information to be included directly in this documentation page.  I've attempted to phrase Dong Cao's words so they fit into the context of this documentation.