Update columnstore-indexes-overview.md
 there is an error in
Convert a columnstore table to a rowstore. CREATE COLUMNSTORE INDEX (Transact-SQL) Usually this conversion isn't necessary, but there can be times when you need to convert. Examples show how to convert a columnstore to a heap or clustered index.

to revert to a rowstore CI we need "create clustered index on.." , to revert it to an heap simply drop CCI the index