Update columnstore-indexes-design-guidance.md
there is an error in
Convert a columnstore table to a rowstore. CREATE COLUMNSTORE INDEX (Transact-SQL) Usually this conversion isn't necessary, but there can be times when you need to convert. Examples show how to convert a columnstore to a heap or clustered index.

to revert to a rowstore CI we need "create clustered index on..WITH DROP_EXISTING=ON" , to revert it to an heap simply drop the CCI index
Is it possible to link directly the right section of the linked document when referring to the md file?
the right section to be referred is ../../t-sql/statements/create-columnstore-index-transact-sql?view=sql-server-2017#e-convert-a-columnstore-table-to-a-rowstore-table-with-a-clustered-index