Correct copy and pasted query
This section of the documentation discussed disabling Managed Backup as a whole, at the Instance Level.  Specifically, the text in this sections paragraph states the need to set the database_name parameter to NULL; however, the SQL Syntax was copy and pasted from above in this document and was not updated to reflect this needed parameter value.  This update is to show this parameter being executed as NULL, as to apply to the whole SQL Instance.