Script to create role and permissions
The instructions, although comprehensive, are SSMS specific. This isn't much use to those using other tools, such as SOS, sqlcmd and mssql-cli, etc. The above also requires a lot of button clicking, which isn't quick. A script is a far quicker way of doing this, and much easier of doing this (and non environment specific) it should be provided.