Update create-an-entity-master-data-services.md
Added a Tip, which was not specified in documentation earlier. In many cases, the Staging table names are not same as Entity name.
Text Added : 
If using the default naming for Staging tables, MDS will automatically append identifiers (e.g. _1, _2) to the staging table names if an entity with same name exists in another Model.