Fix illegal chars, add some info, other editing
I am fixing 4 illegal characters (> 0x7F): curly apostrophes, n-dash.
.
My attention was brought to this article by my scan, for proper Versioning, of all articles that have been directly updated in upstream branch 'release-sqlseattle'.
.
I was surprised to see stub articles in our formal release-sql seattle branch, which will be going live in a few weeks. It might be cleaner to have stub and other incomplete articles instead in a non-release upstream branch, until they are completed and are thus ready for release-sqlseattle. Perhaps opinions vary on the approach. When I encountered the stubs, I decided to look through more PolyBase .md files, which led me to this file.
.
I could find no reason why this article needs to be updated directly in release-sqlseattle, or in other words, why it would need the new Versioning moniker 'sql-server-ver15'.
.
But when my attention was brought to this article, I could not help but notice a few minor editing issues. For example, the first mention of Transact-SQL must be spelled out that way, and only then can later occurrences abbreviate to T-SQL. And the occurrence here of lowercase t-sql is never proper. I notice a couple other minor issues, like the lack of any space between the end of one sentence and the start of the next. Also, the Microsoft style guide explicitly forbids the slash construct of "import/export", partly because the construct harms machine translation into non-English languages.
.
I have been curious about PolyBase, and felt that parts of the existing draft were not quite giving me all the short introductory info that I wanted. I did a little research elsewhere, and felt compelled to add a little info to the existing draft. But I found it kinda necessary to rephrase or rewrite a paragraph or two to make the added info integrate elegantly. I eventually found myself doing more rewriting that I had planned.
.
So my ask is that you review this PR. Then in a Comment in the PR, @ mention me to say if you want me to Close and discard this PR, or for me to #sign-off and Merge it (into 'master'), or whatever else.
.
Thank you Matthew.
GeneMi (MightyPen)