Added CLR data type limitation
Trying to add a filter expression containing a column of data type hierarchyid to a NONCLUSTERED INDEX yields in the following error message:
Filtered index '...' cannot be created on table '...' because the column '...' in the filter expression is of a CLR data type. Rewrite the filter expression so that it does not include this column.