Pull READ_COMMITTED_SNAPSHOT ON out of note
The "If READ_COMMITTED_SNAPSHOT is set to ON" paragraph appears to have been intended as its own bullet, not a note. This PR fixes that.

Note that this PR does not fix the problem that although the page's markdown renders correctly in GitHub, its indentations render incorrectly on docs.microsoft.com. This is a problem with several pages - any time a paragraph follows a bullet. For this page, the problem first shows for the "READ COMMITTED" section. The section is indented, even though it shouldn't be. The incorrect indenting makes the structure of the page difficult for the reader to follow.