Update create-type-transact-sql.md
The comment before the 2 examples appears to be wrong. it's not the case of "disk-based table" vs "memory-optimized table".
the 2 examples show 
1) how to define a user-defined datatype 
2) how to create a table-valued parameter type using the AS TABLE option. in this case you have 
<table_option> ::=  
{  
    [MEMORY_OPTIMIZED = {ON | OFF}]  
}  
Indicates whether the table type is memory optimized. This option is off by default; the table (type) is not a memory optimized table (type). 
So i propose the new comments . Thanks