Missing GO statement in example O
The example is missing a GO statement after the END of the intended procedure definition.
This causes an infinite loop because the EXEC statement is accidentally included in the procedure body.

This could be indicative of a wider problem with the syntax documentation. It states
CREATE PROCEDURE ... AS { [ BEGIN ] sql_statement [;] [ ...n ] [ END ] }  
which implies that if you put a BEGIN immediately after the AS, any statements after the END will not be in the procedure body.
This isn't the case at all - SQL Server must be considering the BEGIN/END as being within sql_statement

You can do
CREATE PROCEDURE dbo.test2
AS
BEGIN
    SELECT 1
END
;

BEGIN
    SELECT 2
END

GO

EXEC dbo.test2

Will print 1 then 2.