Update sys-dm-exec-background-job-queue-stats-transact-sql.md
if we want to know how many failed attempts we got we must compare the set of "all the attempts".  but  the column "enqueued_count" is not the number of "all the attempts" but "Number of requests successfully posted to the queue", so to me it seems that if we want to calculate the set of "all the attempts" we must sum enqueued_count + enqueue_failed_full_count + enqueue_failed_duplicate_count (Number of requests successfully posted to the queue + Number of failed enqueue attempts because the queue is full + Number of duplicate enqueue attempts ) but actually i'm not sure about "Number of duplicate enqueue attempts" (these are to count or not in the set of all the attempts?). maybe a duplicated event raises only when the queue is full so we must exclude the  "Number of duplicate enqueue attempts"  from the set  "all the attempts".   thank you best regards