Update string-split-transact-sql.md
I added a note in the remarks section discussing the order of the result set, which is not currently specified.  I give two options for the new sentence, depending on whether the function returns the results in order or not as indicated by the options in braces.  At the time of writing I do not know which version is correct but it should be stated one way or the other.