Update error-procedure-transact-sql.md
Removed confusing text relating to nested CATCH blocks.

The text referred to the error number, which is dependent on which `CATCH` block is executing, but which is not returned by `ERROR_PROCEDURE`. Also, `ERROR_PROCEDURE` should return the same value in an outer or inner `CATCH` block, so this text is not relevant to documentation of this function.