translate-transact-sql.md - Spelling, Grammar, Wording
The document was quite poorly worded. It contained spelling errors, and read like broken English. It also contained statements that weren't true; if NULL is passed for any parameter, then NULL is returned, however the document states the opposite. The word "should" in the document implies that something isn't always the case. For example "he behavior of the TRANSLATE function should be identical to the REPLACE function. " If TRANSLATE doesn't always operate like REPLACE, when does it not? It's not that TRANSLATE should, it does. I've also removed the repetition of this statement.