TRANSLATE vs REPLACE
I realised, after making a now accepted change, that actually `TRANSLATE` and multiple `REPLACE` statements are not equivalent. If a character is replaced to one that is also appears later in the translatations character string, the character will not be replaced. For `REPLACE`, however, this is not true. `REPLACE` will replace all instances of that character, including those that were replaced in one of the nested functions. I have therefore added this information in the Remarks section.