minor additions, major updates
Added several packages for rhel that are required. Added information about the hostname of the server requirement to match service definition and be resolvable in dns. Added ntpdate time sync information so customers won't fail for skewed clocks. Added notes on keytab file so customers can remove correct entries and it'll be clearer. Removed redundant name server configuration on network interfaces, best practices for name server configuration is that they only be set in resolv.conf.