Update ssis-catalog.md (#1423)
* Update ssis-catalog.md

My organization does not have Management Studio installed on all SQL servers. So, the instructions as provided for starting the Upgrade Wizard do not always apply to us. After searching in Program Files, I found the wizard. It might be useful for others as well.

* Update ssis-catalog.md

Moved the text to the existing sentence/paragraph in the markdown and replaced 'executing' by 'running'