Update configure-http-access-to-analysis-services-on-iis-8-0.md
Following this tutorial, had an 500 - Internal server error caused by not using the previous version of the .dll. 
In my case, it was from the 2012 to the 2017 but it was observed to be the same from 2016 version to 2017. Answer found here: https://social.technet.microsoft.com/Forums/en-US/8960e951-0e77-41a2-8225-863ac0c53b69/through-iis-manager-not-able-to-connect-to-analysis-services-in-2017-version?forum=sqlanalysisservices