It is not necesary to declare the 'http://www.w3.org/2001/XMLSchema-instance'
Apparently the xpath processor includes the http://www.w3.org/2001/XMLSchema-instance by default if not specified with the [WITH XMLNAMESPACES] clause. As long as the namespace is declared in the xml itself an implicit [WITH XMLNAMESPACES('http://www.w3.org/2001/XMLSchema-instance' AS xsi] will be issued.

For example:
DECLARE @doc as XML = CAST ('<?xml version="1.0"?>
<note xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
<name xsi:nil="true"> </name>
<text> hello </text>
</note>
' AS XML);

SELECT @doc.query('note/name[@xsi:nil="true"]') as result;
Will return results even when the xsi namespace is not explicitly declared.

<name xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:nil="true" />

Changing the xml to not mach the standard namespace exactly causes it to not return results.

DECLARE @doc as XML = CAST ('<?xml version="1.0"?>
<note xmlns:xsi="http://www.w3.org/2001/XMLSchema-instanceZZZZ">
<name xsi:nil="true"> </name>
<text> hello </text>
</note>
' AS XML);

SELECT @doc.query('note/name[@xsi:nil="true"]') as result;

This is not true for all xml processors and seems like a design decision. I tested similar scenarios using the .net processors and adding the namespace to the manager is always required.

I am trying to follow the same succinct style for the documentation, although I believe examples are required to better illustrate all these behaviors.

Thank you.