fixing links which hardcoded pointed to the GitHug page instead of rightplacein the document
The links in this page which should point to anchor are pointing to the GitHub Page. They should be relative path and not as bellow example::
https://github.com/MicrosoftDocs/sql-docs/blob/live/docs/t-sql/functions/columns-updated-transact-sql.md#a-using-columns_updated-to-test-the-first-eight-columns-of-a-table

This link take the readers from the Doc to this project instead of redirect him to the right place in the same doc page. The right link should be relative link like bellow:
#a-using-columns_updated-to-test-the-first-eight-columns-of-a-table