Added extra final step
The tutorial gets you to create two windows, One has AE enabled, one doesn't, But you only ever try the code in the one with it enabled. You need to also try it in the other window to see what happens, or it misses a key point of the tutorial.