Description copied from other events is erroneous
Although the description matches that of the StartTime for other events, in this case the StartTime is _not_ the start time of the Audit Logout event, as is the usual case for most other events. It it a reference to the StartTime in the Audit Login event, allowing it to be used to match this Audit Logout event with the corresponding Audit Login event.