This document can be updated with images wherever required so is would be more easy to understand.
I have a white paper with all images added to this document from Microsoft Docs. It makes it easy for user to understand if images from SQL Server management Studio are added here. 
Also, 'Export a bacpac' document is not present with us here. I have a document for that too. 
I am willing to help improve this document and add the additional content for 'Export bacpac'. Please reach out to me to talk more because I am new to Github to edit it myself.