Update scale-out-support-for-high-availability.md
I thought this would be helpful to prevent having to recreate certificates. I wrongly assumed this was the availability group endpoint when I initially configured the master server. It was only when I got to step 7 that I realised that the master service endpoint is a separate computer object with a unique IP address. 

Incidentally, it appears you can just use the availability group endpoint for the master service to get around this. Presumably that reduces some level of HA though, as the endpoint will be abstracted from the clustered service itself.