Update define-colors-on-a-chart-using-a-palette-report-builder-and-ssrs.md
Dear MS, this article says "Custom palettes are report-specific." It looks as if you cannot reuse a palette within a report; it looks like it is chart specific rather than report-specific. Please either correct it or give guidance on how to reuse it for other charts within the report. Thank you.