Update create-login-transact-sql.md
To Julie at Microsoft: (1) I added some additional transaction verbiage that is also a logical consequence, put the content in the header so it will apply to all versions. (2) Microsoft's business model should NOT be to insist that its customers fix the documentation when problems are pointed out. That is what good technical writers are for! I'm losing considerable $$$ per hour when doing free technical writing for Microsoft rather than working for my client.