ALTER CERTIFICATE supports BINARZ private key
Hi there,

we recently had a case where we wanted to transfer a private key from one database to the other, but could not drop the certificate without previously dropping some other things. Turns out, that ALTER CERTIFICATE supports the same BINARY syntax as CREATE CERTIFICATE, it is just not documented. I want to fix this with this pull-request, but I am not 100% sure about all details. Feel free to polish as needed or get back to me so I can improve it myself.

Best Regards