Impacts of CDC when patching SQL server
Add a topic about impacts of CDC when patching instance. If instance enters on script upgrade mode, it will runs proc sp_vupgrade_replication, which runs sp_cdc_vupgrade. In this proc, it can rebuild indexes. I wokred on company that experinced this problem in SQL Server 2008 R2, and after talking with support, it was fixed, but current code (checked on version 2016), can trigger these rebuilds yet. BEcause this, i guess is important alert readers for they plan instance patching.