Add a note that Microsoft uses the in place upgrade approach on their own Azure IAAS production databases
I took the time to suggest this edit because I'm encountering a lot of resistance to using in place upgrades for our production databases.  The people who are resistant point to this documentation which seems pretty out of date to me in light of the blog post 5 years ago indicating that Microsoft have switched to using the in place migration approach.