Added a link to the SQL Server Logs Doc
This doc mentions checking the "error logs" for more info but never mentions how to get to to the logs or which logs specifically should be looked at. I added a link to  https://docs.microsoft.com/en-us/sql/relational-databases/performance/view-the-sql-server-error-log-sql-server-management-studio?view=sql-server-2017

I think if this doc included more info on troubleshooting errors with DAC, it would be helpful.