visually separate key name from other text
The formatting in this commit is only a proposal. The main concern is the name of the key is difficult to parse easily from the text as there's no visual delimiter of what's the key and what's part of the explaination.

If this type of formatting isn't appropriate, feel free to close.