Add Known Issue: Default Language Incorrect
This is a known issue in SSMS 18.2, which I discovered at the following feedback page:

https://feedback.azure.com/forums/908035-sql-server/suggestions/38236363-ssms-18-2-default-language-display-bug

Since it has been just over a month since the issue was acknowledged by Matteo Taveggia, I thought it would be helpful to include the verbiage in the release notes, potentially saving others from the several hours of research (and low-grade panic; more a sense of disquiet, really) that I went through to discover that it was already a known issue. Randy Hartwig's point in the feedback post was well-made.