AllowedResourceExtensionsForUpload property needs to be removed for SSRS 2016 version of doc
The server property named AllowedResourceExtensionsForUpload does not exist for Reporting Services 2016 version and the product team stated that they will not be updating SSRS 2016 to include the property. So we either need to remove the AllowedResourceExtensionsForUpload  property from the SSRS 2016 version of the documentation or update the documentation and state that the property does not exist for SSRS 2016.