Update "float and real" conversion notes
Fixes #2827 

The behavior changed starting in SQL Server 2016. Please see my comments in #2827 for the test code.

Please check the include. The `CREATE TABLE` page uses  `[!INCLUDE[ssSQL15](../../includes/sssql15-md.md)]` but that is a bad name for the include file since there is no 2015 version. There is an identical include file with the correct name, **sssql16**, so I am using that instead. But, I have no way to test include files to hopefully there is no problem with the "16" file.