[Relational databases] Update sql syntax to enable polybase
The RECONFIGURE [ WITH OVERRIDE ]  statement when copied and pasted will fail with a syntax error. Should be either RECONFIGURE; or RECONFIGURE WITH OVERRIDE; I have updated it to RECONFIGURE; as I don't think the OVERRIDE is required. I have also removed the comment "The instance then must be restarted." Won't do any harm, but should not be necessary as in sys.configurations the "is_dynamic" value for "polybase enabled" is set to 1.