Table on page has a set width somewhere and causing a horizontal scrollbar
The main table on this page is not displaying correctly. The content in the first column is too big for the table width and so it is causing a horizontal scroll bar on the bottom of the table. This is not a very good experience for customers to view the page content.