LOCBLOCKER L136588: removing empty file to avoid localization issues
Localization team has reported source content issue that causes localizable text to be blocked from localization.
Description: This file is empty and causing issues in the localization process. The previous approach implemented https://github.com/MicrosoftDocs/sql-docs-pr/pull/12175/files was not enough to fix it, so after discussing with the IPM , we need to remove this file from the source repo, and update/remove all their references in the Articles.